fastapi-starter-boilerplate/requirements_rag.txt
# Core FastAPI dependencies
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
python-multipart==0.0.20

# Database
motor==3.7.1
pymongo==4.13.2

# Authentication
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.3.0

# External APIs
twilio==9.6.5
google-generativeai==0.8.5
httpx==0.28.1

# RAG and Vector Database
chromadb==0.5.23
sentence-transformers==0.3.3
langchain==0.3.12
langchain-community==0.3.12
langchain-chroma==0.1.4

# Additional utilities
pydantic==2.11.7
requests==2.32.4
numpy==2.2.6
